backend/requirements.txt
# Core dependencies
openai-chatkit>=1.1.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.54.0
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0
sqlalchemy>=2.0.0

# Additional utilities
python-multipart>=0.0.6
